test(routines): split the 2,249-line commands/routines.test.ts into 4 parallel slices (RUSH-2819)
routines.test.ts keeps the device pin/activation residual (13 tests); add (18),
list (19), and run (28) move to routines.add.test.ts, routines.list.test.ts,
routines.run.test.ts, sharing routines.test-fixture.ts. All 78 tests preserved
verbatim (vitest-verified twice; the 88 in an earlier comment was a grep
overcount). Measured: 196.84s serial -> 47.14s across parallel forks. The
scheduling group pins add/list via tests: since they drive the CLI without
statically importing routines.ts.
